Walls and Floors
Set all StaticID 31 walls (Granite pillars) to be indestructible. If there were any that were expected to be breakable, they will need to be re-fixed.

Then, made a pass and changed any Floor\Water\Icecomponents that were under an indestructible wall into static components. Unsure if this will _help_ anything, but I expect that consistency of not having floor-type tiles that can never be navigated to will be beneficial going forward.