World forge (#704)
* Caching for offline access

Added a simple 'save to file' for both ComponentResource xml and ScriptingData bytes. If the server connection fails, load those and if successful, continue running.

* Filters

Created a StaticFilter, which only shows Statics. Highlights false floor statics that don't have anything preventing a player from walking on them. Also tweaked Wall filter so that it highlights indestructible walls over FloorComponents and destructible walls without a floor-type component.

* Update caching

Using a ".storage" folder to prevent cluttering the main directory.
Added a message box for when connection fails if the offline mode should be used.

* Only resolve FloorComponent

Since Ice and Water are types of Floor, we don't need to create a separate check. Only if we intended to color them differently.

* Revert "Update caching"

This reverts commit 715c0776cf5396d3624ed17aeaca19642f83b2f3.

* Revert "Revert "Update caching""

This reverts commit 94a279f8d14ad8fa49de85272d2b22fc7b57be58.

* Load cache before init ApplicationWindow

the ComponentsResource gets used when the ApplicationWindow inits, and needs to be present then in order to build the floor, water, structural, etc, options in the Component panel.

Co-authored-by: Jalpesh Kachhadia <jalpeshk@gmail.com>