Jump handlers for non-spawn documents
Entity, Location, Subregion, Treasure and Spawn all now have functions that swap to the correct entry if passed a valid one when sending the 'swap to document' message.
Locations and Subregions got the same treatment Spawns did and will jump-to the right entry if the current world render has one under the selection.