Tweaked Tool and Filter tab dependency
Now that there are renders on more than region tabs, I removed the requirement for the current tab to be a region in order to change the filter.
I also added the ComponentPanel as a valid Activedocument to allow tool changes. This fixes the issue where tools seemed to get 'stuck' because the user selected a new static, then wanted to change tool.