World forge updates (#806)
* Teleporter filter

filter only teleporter components, junk 'floor' icon for now, also highlights teleporters with a 0 for destinationX or destinationY.

* Spawns Document world renders and scaling

Added renders to the Spawns Document with highlighting of the location or regions for spawners.
Modified scaling so that +/- now scale sprites and rectangles correctly.
Added a * hotkey to world renders to display the x10 grid, like Locations had, but available everywhere.

* added portal icon to teleportfilter

* destination highlights - first pass

* destination highlights - multiple rows

destination highlights now show more than one source as separate rows. zoom in if they are unreadable. A good example is Oak undead, as there are a ton of stacked air hexes.

* source highlights, first pass

teleporters have a highlight that indicates where the port to.

like destination highlights, this needs to be converted into a toggle.

* Update TreasuresDocument.xaml

Bound treaures defintion in a grid instead of a stackpanel so you can scroll when there's more than a screen worth of items.

* selected teleporter highlighting

Teleporters or destination tiles cause the highlights to be a different color if contained in the current selection.

* Icon for teleporterfilter

* jump-to tab first pass

Adds some hotkeys to jump to specific documents - hotkeys currently only attached to the render screens and need to be moved higher up the application tree.
also testing adding a hotlink to the entity by clicking a row in the spawns tab.

* MouseWheel zooming

More intuitive

* Added a new render filter to show all spawns

Also cleaned up the 'get a bounding box' code to be less repetitive.

* spawn hotlinking, pass 1

Using the spawn hotkey, Ctrl-P while having a tile containing a locationspawner selected, the spawndocument will jump to the correct entry.

* swap to spawn tab via messager

next, to implement the rest of the tab jumps

* fixed spawner jumping

jump wasn't working if there was no selection.. changed order of things.

* Entity jumping, pass 1

reimplement jumping to entity document as a message

* convert remaining jumps to messages

ApplicationPresenter  now listens for a type to be sent and jumps to the right document for that type. For some types, it might do extra work, like Spawners.

* disabling entity jump from spawns

* spider icon for spawns filter

* Jump handlers for non-spawn documents

Entity, Location, Subregion, Treasure and Spawn all now have functions that swap to the correct entry if passed a valid one when sending the 'swap to document' message.
Locations and Subregions got the same treatment Spawns did and will jump-to the right entry if the current world render has one under the selection.

* Ctrl-Left and Ctrl-Right hotkeys to jump between regions

Used to cleanly cycle between available regions when doing visual inspection of things like spawns or teleporters.

* Tweaked Tool and Filter tab dependency

Now that there are renders on more than region tabs, I removed the requirement for the current tab to be a region in order to change the filter.
I also added the ComponentPanel as a valid Activedocument to allow tool changes. This fixes the issue where tools seemed to get 'stuck' because the user selected a new static, then wanted to change tool.

* Backspace hotkey to jump back to previous tab

only works if current tab is world render for now.

* VisibilityOptions

Co-authored-by: Jalpesh Kachhadia <jalpeshk@gmail.com>