Tile fixes and a few spawn edits (#808)
* Tile fixes and a few spawn edits

Fixed a few places that got missed in previous pass. statics under secret doors, floors under indestructible walls, etc.
Also changed the Leng bushes to static instead of a tile with two floors. Updated the other floor to have the 2 movement cost instead. this was 30-ish replacements across the regions.
Updated Yasnaki level spawns so that they didn't spawn in Shelob's lair, and had a distinct spawner for the area between Shelob and Carfel.
in Oak, I gave the thief trainer an exclusion region and set the thief portal hexes to affect NPCs in order to keep the wildlife out.

* secret door with wrong door statics

I FOUND IT. Someone mentioned this door and I couldn't find their post later... but I spotted the door and now it's fixed.

* fixed borked ice tile