Kesmai loot and balance pass test #1 (#767)
Changed surface crits ( not animals since they shouldnt loot anyway ) and crits on -1 and -2 to CanLoot = false.

Last change for gem test.. Taking a native gem ( in this case the Pear Diamond ) and making it very rare but very valuable in each dungeon level and on each creature.

I started with the chance at 1.5% on Kes -1 with a value of 10,000 and each level increased it be +.3% and +2,000. Just don't know how often these will drop until getting in an testing. May or may not need changing.

This is more or less a look at a possible proof of concept on handling gems in each land. You will have semi rare normal gems that are valuable and 1 native gem that is super rare but very valuable. Kinda like the OG UK Bars but expanded upon.

Added the Shining Steel Longsword to Daisy. It is the same one as the NPC Fighters use in -4. The Shining Steel Greatsword is still missing from item list so we can't add it yet.