Vendors in Leng, Axe, Oak, UK (#846)
* Vendors now have stock

Populated shops in Leng, Axe, Oak and UK.
Each vendor picks a few items from a common stock per vendor class, and then gets a chance of a few rare items.
Leng has a fun new vendor. Go visit Gill, East of the balm and recall vendors and try your luck.

* typo

Caught this typo in Axe, but forgot I had used that as a template elsewhere.

* Por's pub!

I don't remember this place. But the barkeep is named and I added a pawnshop, balm and recall vendor.

* Update Syntax

Co-authored-by: Jalpesh Kachhadia <jalpeshk@gmail.com>