Overhauled keybindings
Moved document swapping keybindings to the applicationpresenter instead of the render screen. This avoids the need for the WorldGraphicsScreen to send a message to the presenter and allows keybinds to be used on tabs without a render.
Fixed the Ctrl-S keybind. It was getting eaten by the S in the WASD map movement handling on the WorldGraphicsScreen.
Added the ability to right-click the SpawnEntry grid on a spawner tab and jump directly to that entity. Alternatively, using Ctrl-E when that datagrid has a selected item will jump to the selected Entity.