World forge improvements (#868)
* Migrate Spawn and Teleport filters to VisibilityOptions

* Better handle Ctrl-input

When Ctrl is held down, parse those bindings and don't fall through to the ones that don't specifically call for Control.

Prevents map from being moved down when pressing Ctrl-S to jump to Segment, for example.

* SpawnsDocument fixes

SpawnsDocument now manually initializes both of its presenters so that they respond when first opening the tab.
Also prevented the 'show spawns' visibility option from working on the spawns document

* Show\Hide Components window when DrawTool is selected

* Floor-type components overwrite by default

Floor, Water, Ice, Air all overwrite eachother instead of appending by default. If you WANT a floor and a water and an air component, hold shift to append.

* Ctrl-D 'jump to destination' shortcut

Selecting a teleporter and hitting Ctrl-D changes the current document to the target region, select the destination hex and (attempts to) centers the screen on it.
Camera movement doesn't work if the destination region has not yet been viewed to initialize it's controls.

* Slightly less terrible icons

Transparency, yay!

* Fixed Erase and Hammer tool interaction with zooming

ComponentTools needed to be adjusted to use the zoomfactor when identifying the current component and rendering it.

* Improved Show\Hide of ComponentsWindow

Moved sending the ToolStart and ToolStop messages to the parent Tool class, and ensured all tools called its onActivate and onDeactivate methods. Added showing the components panel to the Paint tool.

* Show Secret Doors visibility option

Actually implemented earlier, but now has a nicer icon.

* Fixed DrawTool interaction with zooming

DrawTool was also using the old code that didn't reference  zoomfactor.

* Display RegionSpawner stats

Added a few calculated stats to region spawners: average mobs, open floor tiles, density. These can be used to approximate the zoo-ey-ness of a spawner.
Compare Shid Stables, at 40-ish mobs in a single screen at .82 density against most of Kesmai dungeons, which sit around .1.
These stats don't take into account things like how sparse the open floors are and how walls or other obstructions might separate out mobs.

I also tweaked the heights of a few components in the SpawnsDocument. The property grid needed more height on the RegionSpawner tab in order to show the new rows, and places like the LocationSpawner Entries DataGrid, which will probably only ever have a single entry didn't need as much space.

* Overhauled keybindings

Moved document swapping keybindings to the applicationpresenter instead of the render screen. This avoids the need for the WorldGraphicsScreen to send a message to the presenter and allows keybinds to be used on tabs without a render.
Fixed the Ctrl-S keybind. It was getting eaten by the S in the WASD map movement handling on the WorldGraphicsScreen.
Added the ability to right-click the SpawnEntry grid on a spawner tab and jump directly to that entity. Alternatively, using Ctrl-E when that datagrid has a selected item will jump to the selected Entity.

* Fix Entity jump

Null exception if the Spawns document hasn't registered its event before hitting Ctrl-E

* Fix Ctrl-D

Forgot the 'set selection' part when remaking my keybindings and found an edge case where the worldscreen may not be initialized in time if the destination region's tab is in the dockmanager's overflow.

* Entity->Spawn Crossreference

Added a listing of the Spawners that use an Entity definition on the EntityDocument. That list can be used with Ctrl-P to jump to the selected spawner.