Treasture\Entity crossreference with jump functionality
On the Treasures tab, you can now see a listing of Entities that reference that Treasure definition. right-clicking on an Entity offers am action to jump to that object on the Entity tab. Also added right-click-to-jump to the Entities 'Spawns In' listing. Finally, tweaked existing 'ask the spawns document for a current entity' code to always return null if the Spawns document is not the active document.