Total XP and Total HP to Spawner stats
Attempt to parse Entity definitions and extract HP and XP values. This will fail for any complex mobs. For a spawner where all the mobs were parsed correctly, calculate a Total XP and Total HP for the whole spawn and present that next to the stats for average mobs and density.
If HP or XP values cannot be correctly parsed, the field is shown as empty.