World forge (#1040)
* Painting repects Shift\Alt and clobbers "like" tiles

If painting, the behavior is similar to the draw tool. Painting a floor replaces other floor-like components. Unless Shift is held to append, or Alt to replace everything.

* Delete hotkey behavior respects current filter

pressing Delete now only clears components shown by the current Filter. Also triggers a new render pass to reflect changes immediately.

* Clipboard pasting tweak

Clipboard behavior depends on selection size:
1x1 selection, paste the whole clipboard once.
>1x1 selection, paste the clipboard, tiling to fit the space provided.

* Delete button in Components Editor

Added a delete button to remove unwanted components

* Workaround for Counters not changing AccessDirection correctly

Counters now show their current AccessDirection in the dropdown. When a new direction is selected, it overrides all previous choices. Only the LAST accessDirection picked is written to the mappproj.

* Live-updates from ComponentsWindow

ComponentsWindow now updates tile terrain and re-renders the screen when making changes to components.

* fixed mousewheel scrolling in various lists

Entities, Location, Spawnes, Subregions, Treasures all can now be scrolled with a mousewheel.

* ScrollToItem on Spawns and Entities

When jumping to a spawn or entity, the target object is now scrolled into view on the list.

* Ctrl-T Jump to specific Treasure

When using the Ctrl-T 'jump to Treasure tab' hotkey, if the current tab is Entities, and the current editor has text selected that matches a Treasure name, jump directly to that Treasure entry.

* Treasture\Entity crossreference with jump functionality

On the Treasures tab, you can now see a listing of Entities that reference that Treasure definition. right-clicking on an Entity offers am action to jump to that object on the Entity tab. Also added right-click-to-jump to the Entities 'Spawns In' listing. Finally, tweaked existing 'ask the spawns document for a current entity' code to always return null if the Spawns document is not the active document.

* Total XP and Total HP to Spawner stats

Attempt to parse Entity definitions and extract HP and XP values. This will fail for any complex mobs. For a spawner where all the mobs were parsed correctly, calculate a Total XP and Total HP for the whole spawn and present that next to the stats for average mobs and density.
If HP or XP values cannot be correctly parsed, the field is shown as empty.

* Fix for deleting a selection with empty tiles.

Fixes bug introduced in 48925a7: when the current selection includes a tile without data, the delete action would encounter a NullReferenceException.

* Fixed zoom rendering for sprites with Offsets

Offsets also need to be scaled. Only a few tiles have them, but it's noticeable when they are used.