Improved WorldScreen context menu
Enabled the context menu to be displayed when right-clicking inside a selection. Added menu items for turning a selection into a region spawner or subregion, as well as adding the selection to an existing region spawner. visibility of menu items is based on properties of the current selection and active document.