Fixed crash related to cleanup of old documents
When closing a segment, the documents collection is cleared, but the documents remain in memory and will respond to messages until they get GC'd. This change implements a cleanup event that causes all current documents to unregister for events before they are removed.

This really only impacted people who used Ctrl-P or Ctrl-E to move between spawns and Entities AND ALSO opened multiple segments in a single WorldForge session.