Kesmai Spawn tweaks (#1133)
* Kesmai Surface

Split Near East and East apart so they no longer overlap. Removed a few mobs from Near East to account for the smaller region and added some to East to compensate.
Added a crocodile spawn: with East covering so much water, the rare crocs would frequently be swimming offshore. They are now confined to the swamp for newbies to hunt.

* Fixed Elder.Troll exclusion

The bottom of this exclusion was mis-entered, allowing -1 mobs to spawn inside the lair. Finding a zoo and a boss is an unwelcome surprise.

* Split Kes-3 into 5 zones

Leng portal, Axe portal, central, NE, SE.
Density increases in that order. The portals are relatively safe (.07, .09), but the SE corner of the map has increased density (.13). The old zone-wide density was .09.  Approximately 9 net new mobs on average in the zone.

* Split Kes-4 into 5 zones

Oak hallway, Dragon approach, Central, West, South.
Mostly kept density and mob count the same at 84 mobs and .13 density; with the central zone getting a slight boost to .14 and a slight reduction in Oak Hallway and West.