World forge (#1144)
* Painting repects Shift\Alt and clobbers "like" tiles

If painting, the behavior is similar to the draw tool. Painting a floor replaces other floor-like components. Unless Shift is held to append, or Alt to replace everything.

* Delete hotkey behavior respects current filter

pressing Delete now only clears components shown by the current Filter. Also triggers a new render pass to reflect changes immediately.

* Clipboard pasting tweak

Clipboard behavior depends on selection size:
1x1 selection, paste the whole clipboard once.
>1x1 selection, paste the clipboard, tiling to fit the space provided.

* Delete button in Components Editor

Added a delete button to remove unwanted components

* Workaround for Counters not changing AccessDirection correctly

Counters now show their current AccessDirection in the dropdown. When a new direction is selected, it overrides all previous choices. Only the LAST accessDirection picked is written to the mappproj.

* Live-updates from ComponentsWindow

ComponentsWindow now updates tile terrain and re-renders the screen when making changes to components.

* fixed mousewheel scrolling in various lists

Entities, Location, Spawnes, Subregions, Treasures all can now be scrolled with a mousewheel.

* ScrollToItem on Spawns and Entities

When jumping to a spawn or entity, the target object is now scrolled into view on the list.

* Ctrl-T Jump to specific Treasure

When using the Ctrl-T 'jump to Treasure tab' hotkey, if the current tab is Entities, and the current editor has text selected that matches a Treasure name, jump directly to that Treasure entry.

* Treasture\Entity crossreference with jump functionality

On the Treasures tab, you can now see a listing of Entities that reference that Treasure definition. right-clicking on an Entity offers am action to jump to that object on the Entity tab. Also added right-click-to-jump to the Entities 'Spawns In' listing. Finally, tweaked existing 'ask the spawns document for a current entity' code to always return null if the Spawns document is not the active document.

* Total XP and Total HP to Spawner stats

Attempt to parse Entity definitions and extract HP and XP values. This will fail for any complex mobs. For a spawner where all the mobs were parsed correctly, calculate a Total XP and Total HP for the whole spawn and present that next to the stats for average mobs and density.
If HP or XP values cannot be correctly parsed, the field is shown as empty.

* Fix for deleting a selection with empty tiles.

Fixes bug introduced in 48925a7: when the current selection includes a tile without data, the delete action would encounter a NullReferenceException.

* Fixed zoom rendering for sprites with Offsets

Offsets also need to be scaled. Only a few tiles have them, but it's noticeable when they are used.

* Extend Entity with a Threat value

Entities scrape their OnSpawn() code for skill levels (magic skills are doubled). XP, HP and threat are displayed on the entity list in Spawners.
Doesn't take into account things like weapon damage, custom damage ranges, offensive spells type (but does ignore skill levels for blind, fear, stun).

* Minor fix for spell Threat calculation

Spell skill parameter is a Double. Fixed regex to handle a real number instead of an integer.

* Reordering buttons in Component editor

Added 'Move up' and 'Move down' buttons to the component editor.

* Copy\paste support for ComponentEditor

Component Editor supports Ctrl-C and Ctrl-V. If the clipboard contains a single correctly formatted XML snippet for a component, then it is appended to the tile.
Also, delete action in Component Editor selects the next component instead of restarting at component[0]

* Copy\Paste support for Spawners

Spawns document now has three new toolbar icons for Region and Location spawners. They export the selected spawner to the clipboard as XML, clone the selected spawner, or create a new spawner from XML on the clipboard.

* Option for bypassing the 'work offline' prompt

This one is for me. I restart WF so frequently that the extra click is just too much.
Placing a 'WorkOffline.flag' file in .storage prevents the prompt and just automatically opens, even when Stormhalter cannot be reached - such as when launched in debug mode.

* Copy\Paste support for Entities

Adds a Copy to clipboard and a Paste from clipboard button to Entities listing.

* Fix for backwards boolean

I let autocomplete type some code for me without being careful. It got the idea right, but the test backwards.

* Copy\Paste for Subregions

* Copy\Paste for Locations

* Copy\Paste for Treasures

* Help menu with changelog and link to Wiki

Help menu has a link to the WorldForge article on the Wiki and a changelog popup for people who don't troll the github.

* Converted What's New window to HTML

This improves readability... a little. The WebBrowser control doesn't seem to support internal CSS, but plain HTML is better than plain text. I couldn't find anything better that didn't involve building a custom control or a ton of overhead to render the content.

* Improved WorldScreen context menu

Enabled the context menu to be displayed when right-clicking inside a selection. Added menu items for turning a selection into a region spawner or subregion, as well as adding the selection to an existing region spawner. visibility of menu items is based on properties of the current selection and active document.

* 'Select Destination' button in component editor for Teleporters

The Component Editor shows a 'Select Destination' button when the selected component is a TeleportComponent. Clicking this button overrides the normal right-click context menu and exposes a new pair of options to set the destination or cancel. This should make wiring up stairs MUCH easier.

* Fixed crash related to cleanup of old documents

When closing a segment, the documents collection is cleared, but the documents remain in memory and will respond to messages until they get GC'd. This change implements a cleanup event that causes all current documents to unregister for events before they are removed.

This really only impacted people who used Ctrl-P or Ctrl-E to move between spawns and Entities AND ALSO opened multiple segments in a single WorldForge session.

* Fix for crash on segment load if provided file is not valid xml

Also alerts if the root element is not a segment, in case the document is valid XML.

* Scrollbar on 'Spawns in' list for Entities

Specify the height of the listbox so that it gets a scroll bar instead of just growing off the bottom of the screen.

* Convert HP and XP parsing to CodeAnalysis

replaced Regex search method with code analysis.

* Added 'Choose destination' option to context menu

If a tile has only one teleporter, the right-click context menu has an option to set the destination. This saves the user from opening the component editor, selecting the teleporter, and clicking the action button. Efficient!

* Update AssemblyInfo.cs

Co-authored-by: Jalpesh Kachhadia <jalpeshk@gmail.com>