IFragile interface and random item props
An attempt to implement items breaking upon death due to durability loss via IFragile interface. Test item is a random property spear. Note, I'm not sure if I can just override with a random number, or if the randomization needs to be done in the instantiation so I've implemented half the properties each way to test.