Kesmai Overhaul stage 1
So there's a ton of file changes, but they're mostly infrastructure, no balancing changes yet.

What I've created is a creature table generator and 2 test cases on the surface as far out as I can get.

The purpose of this first change is a PoC to make sure it all works before I get too deep, but also it will allow us to have consistency across crits.  We can then adjust the levels of the creatures instead of individual parameters.