Initial Feature
WF Defaults the Custom Art path to .storage, but this can be changes in .storage\CustomArt.cfg to point instead to a local github repo.
If this path contains a Data\Terrain-External.xml, or a WorldForge\Components.xml, those files are used instead of the ones packaged with the client.
Additionally, if new Terrain-External.xml content references new texture sheets, those can be placed in the folder/subfolders and loaded at runtime.
todo: when loading files from external assets, add (or build a document with additions?) the files to the mgcb file.