UK Gargoyle converted to lurker
Converting the gargoyle to a more appropriate monster, and toning it down to one that spawns every 7-10 minutes instead of 3.5 - 5.

Has high attack value, has poison, but doesn't respawn like crazy and punish people trying to hunt in the zone.

Between the pillar and a single monster this will adequately stop scripting without going overboard like it is now according to feedback.