Bug fix for Gem Price Mutators and correct comments
Exactly the same fix on all appropriate maps.

gemsPriceMutatorMedium was intended to be a modifier of -40% to +nothing. Instead it was doing -40% to + 100%

gemsPriceMutatorAboveAverage was intended to be a modifier of + nothing to +40%. Instead it was doing +100% to +40%

Also update the comments on every mutator to reflect the correct values