Revamping Hummingbird Sword guardian.
This is a slight mix between what originally happened and what happened in lands.

IMO this is a battle someone that has duoed TK should be able to solo (as its only soloable) so shouldnt be "that" hard. Just a check you are not a low level.

I believe I have managed to make it so only one person can enter the room at once. You then have to enter the centre of the room and defeat the guardian who throws WWs and hits you to check your fortitude.

This is what I changed:

1) removed normal portal hex
2) replaced with a special invisible entity that on entering their square, checks if you are holding a hummer amulet And if the guardian is available and alone (no other players)
2.1) If the guardian is not spawned currently then the player is refused entry
2.2) If someone else is already in the guardian room with the guardian then the 2nd player is refused entry. As per the original game this is a solo challenge to show your basic fortitude
2.3) If the above pass then you art teleported in and the amulet is consumed
3) the guardian is only attackable by physical means. No magic and no conc bottles etc. This is meant to be a basic show of fortitude.
4) The guardian will only start attacking when you enter the centre of the room. From that point it will be able to move and cast etc.
5) Guardian has a weak WW but its instant cast and will cast 2 in quick succession for effect. The combined effect will be a reasonable tick of damage but wont last too long. The attacks then are again reasonable but not too bad.