Adding a new Test Region
Adding a new Test Region currently with nothing in it, with the intention of trying new features in a more controlled environment before putting them live. It won't stop compile errors but if someone wants to try some new code and check it works before.. say, breaking the hummer portal, they could test it here first.

Each test could be put through as a separate commit so easy to just roll back to this empty state.