oak -1 tweaks
poison was a little too harsh for this level. bumped exp slightly for certain poison creatures as they also dont give loot so needs to be a good reason for risk/reward.

Also rebalanced the spawn locations to try to make less of a zoo as you come down and also maybe slightly less snakes and scorps.