Oak 2 and 1 tweak (#1682)
* oak -1 tweaks

poison was a little too harsh for this level. bumped exp slightly for certain poison creatures as they also dont give loot so needs to be a good reason for risk/reward.

Also rebalanced the spawn locations to try to make less of a zoo as you come down and also maybe slightly less snakes and scorps.

* oak -2

I merged back in my -2 changes. bumping it up a tiny bit. Big overhaul to the spawn locations. Also added in an extra secret door