Update Blackburrow.mapproj
When the item equips, and unequips, a constant value will be removed. We shouldn't need to save that value.

If the value could not be recalculated at Unequip, then we should store the value and remove that.

Since the values will recalculate each time, they don't need to be serialized.  We should serialize data that needs to be reconstructed at loading.  Since the critbonus and damage are dependent on Quality, we can recalculate them.

Serialized data should be something that we can not "reconstruct" at loading.