Update Blackburrow.mapproj
We don't need to add bonus damage to the item as we equip and unequip it. It can always just be constant.    We could do it on Equip and Unequip -- if we wanted to for example, do it only for a specific class.  Then the original implementation is what you would do, minus the serialization.

You would serialize those properties if, for example, the item spawned with random bonus damage -- you would want to be able to reconstruct that when the item loads.