Update Blackburrow.mapproj
We always return 5, plus a variable that gets modified during Equip/Unequip.  Unequip always just sets the value back to 0.

This is an example of where we can cache some data, but it doesn't need to be serialized because it is dependent on Quality and can be recalculated after loading.