Base.Nod.2025.06.28a (#2254)
* Adding prefab functionality and paint tool adjustment

This change adds a few upgrades to the overall tool.

Allows for the creation of custom prefabs based upon built tiles in an existing segment. To create a prefab user can inspect the tile and save as prefab with a name. Afterword that prefab will be located in the bottom of the dropdown list for components.

Additionally the paint tool will now allow the user to paint custom tiles by popping up at the tool usage to set the configurations.

* Update EntitiesDocument.xaml.cs

* Copy clipboard crash fix

---------

Co-authored-by: Tamarand247 <Tamarand247@gmail.com>